Refactor telegram bot into modules
